fixed data

+++ src/main/java/com/example/muslis/repositories/AlbumRequestRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlbumRequestRepository extends JpaRepository<AlbumRequest, Integer> {
+public interface AlbumRequestRepository extends JpaRepository<AlbumRequest, Long> {
 }
